Align user menu styling with app branding
